Reduce thresholds and add new LandmarkDistance trigger
Thresholds for specific actions have been reduced to improve responsiveness. In addition, a new LandmarkDistance trigger has been introduced for 'mouth small' with specific distance thresholds and designated direction. This aim of this modification is to allow more precise and flexible control based on landmark distances.

Add docstring and modify LandmarkDistance attributes

Extended the LandmarkDistance class in the ctrlability/triggers module with a descriptive docstring. In the initialization, the timer is now converted to seconds, commentary to "triggered" attribute added and other minor modifications implemented to improve clarity and code maintainability.