Add KeyCommand to ctrlability actions
KeyCommand action is added to the ctrlability actions, enabling triggering keyboard shortcuts as actions. This new class validates that the command is an array of strings. This modification is closely tied with changes in config-expressions.yaml, where the 'Logger' action is replaced by 'KeyCommand' with the command 'w' as an argument upon 'browOuterUpLeft' trigger.