Switch project from requirements.txt to pyproject.toml
The requirements.txt has been deleted and the dependencies have been moved to pyproject.toml. This switch is part of a transition to use Poetry as the package manager. The updates in this commit also include the addition of setup.py and the addition of an "*.egg-info" entry in the .gitignore to exclude Python egg files.